Refactor snapshot UI into StateSnapshots component
